DFPL-1989 Documents added during case creation are not shown in CFV after submission
