+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/CaseData.java
@@ -1107,7 +1107,7 @@ public class CaseData extends CaseDataParent {
         return hearingEndDate.isBefore(LocalDateTime.now()) || hearingStartDate.isBefore(LocalDateTime.now());
     }
 
-    @Deprecated(since = "DFPL-1438")
+    // It will be used in "upload-documents" event (when the case is in Open state)
     private final List<Element<ApplicationDocument>> applicationDocuments;
     private final String applicationDocumentsToFollowReason;
