+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/ApplicationDocumentsService.java
@@ -3,18 +3,32 @@ package uk.gov.hmcts.reform.fpl.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType;
 import uk.gov.hmcts.reform.fpl.model.ApplicationDocument;
+import uk.gov.hmcts.reform.fpl.model.ManagedDocument;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
 import uk.gov.hmcts.reform.fpl.service.time.Time;
 import uk.gov.hmcts.reform.fpl.utils.DocumentUploadHelper;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static java.util.stream.Collectors.toList;
 import static org.apache.commons.lang3.ObjectUtils.isEmpty;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.BIRTH_CERTIFICATE;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.CARE_PLAN;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.CHECKLIST_DOCUMENT;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.GENOGRAM;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.OTHER;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.SOCIAL_WORK_CHRONOLOGY;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.SOCIAL_WORK_STATEMENT;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.SWET;
+import static uk.gov.hmcts.reform.fpl.enums.ApplicationDocumentType.THRESHOLD;
+import static uk.gov.hmcts.reform.fpl.utils.ElementUtils.element;
 import static uk.gov.hmcts.reform.fpl.utils.ElementUtils.findElement;
 
 @Service
@@ -24,6 +38,44 @@ public class ApplicationDocumentsService {
     private final Time time;
     private final DocumentUploadHelper documentUploadHelper;
 
+    private Map<String, Object> synchroniseToNewFields(List<Element<ApplicationDocument>> applicationDocuments,
+                                                       List<ApplicationDocumentType> applicationDocumentTypes,
+                                                       String newFieldName) {
+        final List<Element<ManagedDocument>> newDocListLA =
+            Optional.ofNullable(applicationDocuments).orElse(new ArrayList<>()).stream()
+                .filter(fed -> applicationDocumentTypes.contains(fed.getValue().getDocumentType()))
+                .filter(fed -> fed.getValue().isConfidentialDocument())
+                .map(fed -> element(fed.getId(), ManagedDocument.builder()
+                    .document(fed.getValue().getDocument())
+                    .build()))
+                .collect(toList());
+
+        final List<Element<ManagedDocument>> newDocList =
+            Optional.ofNullable(applicationDocuments).orElse(new ArrayList<>()).stream()
+                .filter(fed -> applicationDocumentTypes.contains(fed.getValue().getDocumentType()))
+                .filter(fed -> !fed.getValue().isConfidentialDocument())
+                .map(fed -> element(fed.getId(), ManagedDocument.builder()
+                    .document(fed.getValue().getDocument())
+                    .build()))
+                .collect(toList());
+
+        Map<String, Object> ret = new HashMap<>();
+        ret.put(newFieldName, newDocList);
+        ret.put(newFieldName + "LA", newDocListLA);
+        return ret;
+    }
+
+    @SuppressWarnings("unchecked")
+    public Map<String, Object> synchroniseToNewFields(List<Element<ApplicationDocument>> applicationDocuments) {
+        Map<String, Object> ret = new HashMap<>();
+        ret.putAll(synchroniseToNewFields(applicationDocuments, List.of(SWET, SOCIAL_WORK_CHRONOLOGY,
+                SOCIAL_WORK_STATEMENT, GENOGRAM, CHECKLIST_DOCUMENT, BIRTH_CERTIFICATE, OTHER),
+            "documentsFiledOnIssueList"));
+        ret.putAll(synchroniseToNewFields(applicationDocuments, List.of(CARE_PLAN), "carePlanList"));
+        ret.putAll(synchroniseToNewFields(applicationDocuments, List.of(THRESHOLD), "thresholdList"));
+        return ret;
+    }
+
     public Map<String, Object> updateApplicationDocuments(List<Element<ApplicationDocument>> currentDocuments,
                                                           List<Element<ApplicationDocument>> previousDocuments) {
         List<Element<ApplicationDocument>> updatedDocuments = setUpdatedByAndDateAndTimeOnDocuments(
@@ -32,6 +84,7 @@ public class ApplicationDocumentsService {
         Map<String, Object> data = new HashMap<>();
 
         data.put("applicationDocuments", updatedDocuments);
+        data.putAll(synchroniseToNewFields(new ArrayList<>(updatedDocuments)));
 
         return data;
     }
